Moved `html.js` to `utils/html.ts`
