+++ src/bandcamp/pages/page-music.js
@@ -1,4 +1,4 @@
-import { contentChangeWithPolling, createDatalistFromArray, createElementFromHTML, input, selectElementWithContent, setDataAttribute } from "../../modules/html.js";
+import { contentChangeWithPolling, createDatalistFromArray, createElementFromHTML, input, selectElementWithContent, setDataAttribute } from "../../utils/html";
 import { containsOneOf, splitString, isEmptyArray, countOccurrences, removeBrackets } from "../../utils/utils";
 import { getBandPhotoSrc, getReleasesData } from "../modules/html.js";
