+++ src/discogs/modules/initialization.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import { click } from "../../modules/html.js";
+import { click } from "../../utils/html";
 import { convertNewlinesToBreaks } from "../../utils/utils";
 import { setSectionHint, fillDurations, getSubmissionFormSectionNotes, selectFormatDescription, selectFormatFileType, setInputValue, getArtistNameInput, getQuantityInput, getNotesTextarea, getSubmissionNotesTextarea } from "./draft-page.js";
 import { showNotificationError, showNotificationInfo } from "./notification.js";
