+++ src/popup/popup.js
@@ -4,7 +4,7 @@ import { loadDiscogsGenres } from "../discogs/modules/genres.js";
 import { loadKeywordMapping } from "../bandcamp/modules/mapping.js";
 import config from "../config.js";
 import { setHistoryTabSearchValue, setupHistoryTab } from "./tabs/history_tab.js";
-import { disable, enable, hide, show, click } from "../modules/html.js";
+import { disable, enable, hide, show, click } from "../utils/html";
 import { setupReleasesTab } from "./tabs/releases_tab.js";
 import { setupReleaseTab } from "./tabs/release_tab.js";
 import { setupCsvDataTab } from "./tabs/csv_data_tab.js";
