+++ src/popup/tabs/download_tab.js
@@ -2,7 +2,7 @@ import { Release } from "../../app/release.js";
 import { DiscogsCsv } from "../../discogs/app/discogs-csv.js";
 import { downloadCsv, objectsToCsv } from "../../utils/csv";
 import { isEmptyArray } from "../../utils/utils";
-import { disable, enable } from "../../modules/html.js";
+import { disable, enable } from "../../utils/html";
 
 /**
  * @param {Element} button
