+++ src/popup/tabs/history_tab.js
@@ -1,7 +1,7 @@
 import { DiscogsCsv } from "../../discogs/app/discogs-csv.js";
 import { ReleasesList } from '../components/releases-list.js';
 import { downloadCsv, objectsToCsv } from "../../utils/csv.ts";
-import { createElementFromHTML, hasDataAttribute, setDataAttribute } from "../../modules/html.js";
+import { createElementFromHTML, hasDataAttribute, setDataAttribute } from "../../utils/html";
 import { clearStorage, clearStorageByKey, findAllReleases, findReleasesByUrls } from "../../modules/storage.js";
 import { populateReleasesList, removeButtonLoadingState, setButtonInLoadingState } from "../helpers.js";
 import config from "../../config.js";
