+++ src/popup/tabs/releases_tab.js
@@ -1,6 +1,6 @@
 import { ReleaseItem } from "../../app/release.js";
 import { getCurrentTab, openTabs } from "../../utils/chrome";
-import { input } from "../../modules/html.js";
+import { input } from "../../utils/html";
 import { findReleasesByUrls } from "../../modules/storage.js";
 import { populateReleasesList, removeButtonLoadingState, setBackgroundImage, setButtonInLoadingState } from "../helpers.js";
 import { downloadReleasesCsv } from "./download_tab.js";
